backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
Markdown==3.8
psycopg==3.2.7
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
django-cors-headers==4.7.0
django-filter==25.1
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
